+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "vscode-copy-from-terminal" extension will be documen
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.0.8
+
+- Shorten initializer command into a shell script invocation
+- Avoid cluttering shell history (e.g., `bash_history`) with initializer command
+
 ## 0.0.5
 
 - Add `tee2code` command